Add ghg block into ghg themes page
